requirements.txt
openai
langchain
pinecone-client
python-dotenv
streamlit
pytesseract
Pillow
pdf2image
tqdm
